LU-12514 tests: add testing for lustre_tgt type
Add test-framework handling for the lustre_tgt filesystem type
using /sbin/mount.lustre_tgt if it is installed on the servers.

Convert the core test-framework.sh code over to call $MOUNT_TGT
instead of $MOUNT_CMD or 'mount -t lustre'. Not all callers
have been converted yet, both to continue testing server mounts
with '-t lustre' as well as reducing the scope of this patch, as
each modified subtest triggers a review-subtest-change session.

Tests still using $MOUNT_CMD or 'mount -t lustre' can be updated
in a follow-on patch.

While mount.lustre_tgt has existed for a long time, installation
was broken in 2.17.0, so it cannot be tested with those servers
during interop sessions.

Fix lustre/utils/Makefile.am to properly install mount.lustre_tgt.
The $(SERVER) env variable was not set in the Makefile, use SERVER.
For Debian servers the $(DESTDIR)$(rootsbindir) was not created in
the install tree before the symlink was being created, do it first.
For RedHat servers using %ghost means the link was never installed.

Test-Parameters: trivial
Fixes: 510aea4a37 ("LU-12514 utils: add "lustre_tgt" filesystem type")
Fixes: 6493d8997e ("LU-17843 build: mount.lustre_tgt as symlink")
TLC-bug-id: TLU-95
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Signed-off-by: Minh Diep <mdiep@thelustrecollective.com>
Change-Id: I5b889b5bbb43475b702baab691deba6d333ebbe5